Recipe DATAIKUOCRDEMO.compute_w3DdXIhY saved through notebook notebook editor for compute_w3DdXIhY

+++ recipes/compute_w3DdXIhY.py
@@ -20,12 +20,14 @@ pdfs_processed_info = pdfs_processed.get_info()
 pdfs_folder_path = pdfs_processed.get_path()
 
 # -------------------------------------------------------------------------------- NOTEBOOK-CELL: CODE
-for output_pdf_folder in os.listdir(pdfs_folder_path):
-    output_pdf_folder_path = os.path.join(pdfs_folder_path, output_pdf_folder)
-    for page in os.listdir(output_pdf_folder_path):
-        page_path = os.path.join(output_pdf_folder_path, page)
-        os.remove(page_path)
-    os.rmdir(output_pdf_folder_path)
+# If project variable is set to true, all outputs cleared and PDF processing is re-run for all input PDFs.
+if dataiku.get_custom_variables()["reprocess_PDFs"].lower() == "true":
+    for output_pdf_folder in os.listdir(pdfs_folder_path):
+        output_pdf_folder_path = os.path.join(pdfs_folder_path, output_pdf_folder)
+        for page in os.listdir(output_pdf_folder_path):
+            page_path = os.path.join(output_pdf_folder_path, page)
+            os.remove(page_path)
+        os.rmdir(output_pdf_folder_path)
 
 # -------------------------------------------------------------------------------- NOTEBOOK-CELL: CODE
 for pdf_folder in os.listdir(images_folder_path):
@@ -33,12 +35,14 @@ for pdf_folder in os.listdir(images_folder_path):
     PDF_pages = os.listdir(os.path.join(images_folder_path, pdf_folder))
     PDF_pages_path = [os.path.join(pdf_folder_path, page) for page in os.listdir(pdf_folder_path)]
     pdfs_processed_path = os.path.join(pdfs_folder_path, pdf_folder)
-    os.mkdir(pdfs_processed_path)
-
-    for i, page_path in enumerate(sorted(PDF_pages_path)):
-        print(page_path)
-        output_path = os.path.join(pdfs_processed_path, 'page_'+"{0:0=2d}".format(i+1)+'.txt')
-        pdf_text = pytesseract.image_to_string(Image.open(page_path))
-        with open(output_path, "w") as f:
-            f.write(pdf_text)
-            f.close()
+    if os.path.exists(pdfs_processed_path):
+        print('PDF already processed.')
+    else:
+        os.mkdir(pdfs_processed_path)
+        for i, page_path in enumerate(sorted(PDF_pages_path)):
+            print(page_path)
+            output_path = os.path.join(pdfs_processed_path, 'page_'+"{0:0=2d}".format(i+1)+'.txt')
+            pdf_text = pytesseract.image_to_string(Image.open(page_path))
+            with open(output_path, "w") as f:
+                f.write(pdf_text)
+                f.close()
